service_paiement/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
pika==1.3.2
requests